Resize scratchpad windows automatically
While it's possible that differently-sized scratchpad windows may
present some advantages, it's much more often the case that a window is
opened that is a sub-optimal size for something centered on the screen.
Automating the resizing of windows as part of the scratchpad toggle
process avoids this entirely and ensures consistent dimensions from app
to app, which is a worthwhile trade-off.